requirements.txt
asgiref==3.8.1
boto3==1.35.92
botocore==1.35.92
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.4
django-auditlog==3.0.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-storages==1.13.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jmespath==1.0.1
lat_lon_parser==1.3.1
packaging==24.2
pillow==11.1.0
praytimes==2.3.2
psycopg2-binary==2.9.10
pulsar-client==3.6.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uWSGI==2.0.26